+++ src/bot/constructors/command.js
@@ -1,5 +1,3 @@
-const chalk = require("chalk");
-
 module.exports = exports = class BananenBase_Command {
   constructor(client, help, check = {}) {
     this.client = client;
@@ -15,13 +13,13 @@ module.exports = exports = class BananenBase_Command {
 
     if (!check.permLevel) this.permLevel = 0;
     else this.permLevel = check.permLevel;
-    if (!check.func) this.check = () => {return true};
+    if (!check.func) this.check = () => {return true;};
     else this.check = check.func;
     if (!check.permissions) check.permissions = {};
     if (!check.permissions.me) check.permissions.me = [];
     if (!check.permissions.user) check.permissions.user = [];
     this.permissions = check.permissions;
   }
-}
+};
 
 exports.event = require("./event/command.js");